youtube-torrent/backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
yt-dlp==2025.9.26
python-dotenv==1.0.1
gunicorn==21.2.0
